% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{age_to_100day}
\alias{age_to_100day}
\title{Convert age to 100 day equivalents}
\usage{
age_to_100day(age)
}
\arguments{
\item{age}{Vector of ages}
}
\value{
Vector of 100 day equivalents
}
\description{
Convert ages into 100 day equivalents using assumed mortality rate. This
function is used internally in \code{\link{get_totals}} and \code{\link{get_distributions}} but it
can be called independently if desired.
}
\details{
Assumes a natural mortality rate of 0.04 per day. Uses the equation \code{exp(-0.04*(100-age))}.
}
\examples{
age_to_100day(c(90,100,110))
}
\keyword{functions}
