% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_generation.R
\name{get_logcpueindex}
\alias{get_logcpueindex}
\title{Get CPUE index}
\usage{
get_logcpueindex(
  df,
  var = "TOTAL_NO",
  group = NULL,
  standardized = TRUE,
  sd = FALSE
)
}
\arguments{
\item{df}{Data frame output from \code{\link{get_totals}}.}

\item{var}{The variable in \code{df} for which to compute the index, typically "TOTAL_NO"
(default), "BIOMASS", or "N100".}

\item{group}{Optional grouping variable in \code{df} for which to compute separate
indices. Typically "STRATA". Multiple grouping variables are allowed.}

\item{standardized}{Whether or not to compute a scaled index, in addition. Defaults
to TRUE. Scaling is done within groups for each NAME, across years.}

\item{sd}{Calculate standard deviations and standard errors. This is done after averaging
repeat tows at the same station, so n is the number of stations. Defaults of FALSE.}
}
\value{
A data frame with columns NAME, YEAR, any grouping variables, and an index.
}
\description{
Given an output table from \code{\link{get_totals}}, calculates the mean log(CPUE+1)
for each YEAR and NAME. Allows an optional grouping variable (e.g. STRATA).
}
\details{
If there are multiple hauls at the same station in a given year, these are
averaged first, and then all stations means in a given year are averaged.

NAs will be filled in for years with no hauls in a given grouping.
}
\examples{
load_trawls()
least_fav_orgs=data.frame(SPECIES=c(2058,2842),
  MATURITY=c("U","U"),
  NAME=c("Pyrosomes","Egg jellies"))
least_fav_orgs_abund=get_totals(speciestable=least_fav_orgs, what="abundance")
least_fav_orgs_index <- get_logcpueindex(least_fav_orgs_abund, var = "TOTAL_NO", group="STRATA")

anchovytable <- data.frame(SPECIES=209, MATURITY="A", NAME="Adult Anchovy")
anchbiomass <- get_totals(anchovytable, what = "biomass")
anchindex_abund <- get_logcpueindex(anchbiomass, var = "TOTAL_NO", group="STRATA")
anchindex_biomass <- get_logcpueindex(anchbiomass, var = "BIOMASS", group="STRATA")

\dontrun{
rockfish100equiv <- get_totals(sptable_rockfish100, what = "100day")
rockfish100index <- get_logcpueindex(rockfish100equiv, var="N100", group="STRATA")
}
}
\keyword{functions}
