% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{get_lw_regression}
\alias{get_lw_regression}
\title{Get length-weight regression}
\usage{
get_lw_regression(species, maturity, plot = F)
}
\arguments{
\item{species}{Species code}

\item{maturity}{Maturity code}

\item{plot}{Logical. Whether to produce a plot (logical). Plots will not be
produced for published regressions, just ones fit using the WEIGHT data.}
}
\value{
A function which takes a vector of lengths (mm) as input, and returns
a vector of masses (g).
}
\description{
Runs a length-weight regression using data in the WEIGHT table and outputs a
function to convert lengths (mm) into weights (g). This function (and the outputed
function) are used internally in \code{\link{get_totals}} and \code{\link{get_distributions}} but
it can be called independently, e.g. if you want to see what the regressions
look like or use them for independent conversions.
}
\details{
This fits a linear regression to log-log data. The outputed conversion
function will do the transformations and backtransformations automatically, so
the inputed lengths should be in the original units (mm).

Length-weight regressions can be done for all species in the WEIGHT table
(run \code{unique(WEIGHT$SPECIES)} for a list) and for all species in
\code{\link{sptable_lw}}.

The rockfishes are divided into 4 groups given in table \code{\link{rflwgroups}}. The
groups are based on general body shape and similarity of regression
intercepts. Within a given group, species with length-weight data are pooled,
and the regression is used for all species in that group, including those
without length-weight data. Thus the species with data act as proxies for
species without data.

For anchovy (209) and Pacific sanddabs (147), adults (A) and juveniles (Y)
are pooled for the regression.

For northern lampfish (661) and bigfin lanternfish (669), data from both
species are pooled for the regression. Unknown myctophids (407) will use
this regression.

For speckled sanddab (148), the regression for pacific sanddab (147) is used.

For octopus (2026), which has weight but no length data, the mean weight is
used.

This function includes and will use published length-weight regressions for
select species not in the WEIGHT table. These species are sardine (562), T.
spinfera (1473, also used by 1791), and E. pacifica (1816, also used by 1847,
as well as 1472, 1846, 2829, 2830, 2835, 2849 assumed to be the mean size of
all measured krill).

If the function is unable to generate a regression for the input species, a
warning will be displayed and the resulting function will return NA for any
length input.
}
\examples{
load_trawls()
anchreg <- get_lw_regression(species = 209, maturity = "A", plot=TRUE)
anchreg(c(120,110,100)) #weights of fish 120, 110, 100 mm

sardreg <- get_lw_regression(species = 562, maturity = "A")
print(sardreg) #this uses a published regression

}
\references{
Sardine regression (OLS) from: Palance, D. G., Macewicz, B. J.,
Stierhoff, K., Demer, D. A., & Zwolinski, J. P. (2019). Length conversions
and mass–length relationships of five forage‐fish species in the California
current ecosystem. Journal of fish biology, 95(4), 1116-1124.

Krill length-dry weight regressions from: Fisher, J.L., Menkel,
J., Copeman, L., Shaw, C.T., Feinberg, L.R. and Peterson, W.T., 2020.
Comparison of condition metrics and lipid content between Euphausia
pacifica and Thysanoessa spinifera in the northern California Current, USA.
Progress in Oceanography, 188, p.102417.

Krill dry-wet weight conversion factor from Williamson, S.W., Favuzzi, J.
and Cox, J.L., 1986. Patchiness and nutritional condition of zooplankton in
the California Current. Fishery Bulletin, 84(1), pp.157-176.
}
\seealso{
\code{\link{sptable_lw}}, \code{\link{rflwgroups}}
}
\keyword{functions}
