% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbers.R
\name{get_numbers}
\alias{get_numbers}
\title{Underlying function for \code{get_totals} and \code{get_distributions}}
\usage{
get_numbers(
  speciestable,
  datasets = "RREAS",
  startyear = 1983,
  what = c("abundance", "biomass", "100day"),
  aggregate = TRUE
)
}
\arguments{
\item{speciestable}{Dataframe containing species information (see details).}

\item{datasets}{Character vector indicating which dataset(s) to use. Multiple dataset can
be specified. Options are "RREAS","ADAMS","PWCC","NWFSC". If unspecified, just uses RREAS.}

\item{startyear}{Start year (default is 1983).}

\item{what}{Either "abundance","biomass", or "100day".}

\item{aggregate}{If TRUE, function will produce total abundance if
what="abundance", total biomass if what="biomass", and 100 day standardized
abundance if what="100day". If FALSE, function will produce size
distribution if what="abundance", mass distribution if what="biomass", and
age distribution if what="100day".}
}
\value{
A single dataframe. See \code{\link{get_totals}} and \code{\link{get_distributions}} for more details.
}
\description{
Underlying master function for get_totals and get_distributions.
It can be called directly, however the other functions may be more intuitive.
}
\details{
The same set of steps is required to obtain all the various distributions and totals.
This function loops through each row of \code{speciestable}, and simply exits the loop at different
points depending on the data requested. The results are stacked, and if totals are requested,
results for entries with the same NAME are added together.
}
\seealso{
\code{\link{get_totals}}, \code{\link{get_distributions}}
}
\keyword{functions}
