% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_depth_stratified_tows.R
\name{load_depth_stratified_tows}
\alias{load_depth_stratified_tows}
\title{Extract RREAS depth-stratified tows}
\usage{
load_depth_stratified_tows()
}
\value{
Table is written to the global environment (and will overwrite any existing tables
with the same names). Diplays "HAULDEPTHSTRATIFED created." if successful.
}
\description{
Pulls out the subset of depth-stratified tows from the HAUL table. Creates the table
HAULDEPTHSTRATIFED.
}
\details{
RREAS standard tows are conducted at 30 m headrope depth (DEPTH_STRATA 2),
with the exception of stations with a bottom depth of less than 60 m, which
are towed at 10 m headrope depth (DEPTH_STRATA 1). These are the tows which
appear in HAULSTANDARD.

Historically, mostly before the coastwide expansion in 2004, multiple depth
strata (DEPTH_STRATA 1: 10 m, DEPTH_STRATA 2: 30 m, DEPTH_STRATA 3: 90 m)
were sampled in succession at specific stations, mostly at stations 110, 125,
133, and 170, but occassionally others. This function pulls out these
depth-stratified tows into the table HAULDEPTHSTRATIFED. It has the same
format as HAULSTANDARD, but with a few extra columns: DEPTH_STRATA, SWEEP (indicates
which of the 3 passes the sampling is from; there is generally one set of
depth-stratified tows per sweep, but not always), and SWEEP_SEP (separates cases
in which there are multiple sets of depth stratified tows per sweep, and sets
of depth stratified tows where SWEEP in NA, which occurs after 2004;
otherwise equal to SWEEP). Each set of consecutive depth stratified tows will
have a unique CRUISE/STATION/SWEEP_SEP value.

HAULDEPTHSTRATIFED can be passed to \code{get_totals} or \code{get_distributions} to
get catch data from these hauls instead of HAULSTANDARD by supplying it under
\code{haultable}.

Depth-stratified tows are defined as valid tows at the same station done in
succession at more than one depth strata. These are cataloged in the internal table
\code{ds_tows}.
}
\examples{
load_trawls()
load_depth_stratified_tows()
anchovytable <- data.frame(SPECIES=209, MATURITY="A",NAME="Adult Anchovy")
anchovyabund <- get_totals(anchovytable, what = "abundance", haultable = HAULDEPTHSTRATIFIED)

}
\keyword{functions}
