% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_erddap}
\alias{load_erddap}
\title{Load RREAS Trawl Data from ERDDAP}
\usage{
load_erddap()
}
\value{
Tables are written to the global environment (and will overwrite any existing tables
with the same names). Diplays "Data loaded" if successful.
}
\description{
Loads the HAUL, CATCH, LENGTH, STATIONS, and SPECIES_CODES tables from the RREAS survey
as currently stored in NOAA's ERDDAP database. These are reformatted as
relational tables to match the format of the tables in the database. This
dataset contains data starting in 1990 and only from standard, active
stations. The function also loads a HAULSTANDARD table with a standardized
set of columns including YEAR, MONTH, JDAY, and lat/lon in decimal degrees.
See \code{\link{RREAS_TABLES}} for metadata. This function has no arguments.
}
\details{
The data tables are stored internally in the package (internet
connection is not required). The maintainers will update the package data
whenever the data on ERDDAP is updated.

As all hauls on ERDDAP are from standard, active stations, HAUL and
HAULSTANDARD will have the same number of rows. HAULSTANDARD will have some
additional columns and the columns will be in a different order. This is just
to match the HAULSTANDARD table generated from the database.

This dataset (unlike the database) contains additional SPECIES/MATURITY
categories 1472/T (total krill) and 1940/T (total rockfish). These export
as NaN from ERDDAP (prior to 2016), but are corrected here.

Abundance (including abundance with size limits) and size distributions can
be obtained from this dataset using \code{\link{get_totals}} and \code{\link{get_distributions}}.

The WEIGHT and AGE tables are not included, so requests for biomass, age
standardized abundance, mass distributions, or age distributions will be
unsuccessful.
}
\examples{
load_erddap()

}
\references{
Sakuma, K.M., Field, J.C., Mantua, N.J., Ralston, S., Marinovic,
B.B. and Carrion, C.N. (2016) Anomalous epipelagic micronekton assemblage patterns
in the neritic waters of the California Current in spring 2015 during a
period of extreme ocean conditions. CalCOFI Rep. 57:163-183
}
\seealso{
\code{\link{load_trawls}}, \code{\link{RREAS_TABLES}}
}
\keyword{functions}
