% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_trawls}
\alias{load_trawls}
\title{Load RREAS Trawl Data}
\usage{
load_trawls(
  activestationsonly = TRUE,
  stdtimeperiodonly = TRUE,
  startyear = 1983
)
}
\arguments{
\item{activestationsonly}{Logical. Include only active stations in HAULSTANDARD. Defaults to TRUE.
If FALSE, the ACTIVE column can always be used to subset later.}

\item{stdtimeperiodonly}{Logical. Include only surveys performed during the standard sampling
time period in HAULSTANDARD (exclude early surveys). Defaults to TRUE.}

\item{startyear}{Start year (default is 1983).}
}
\value{
Tables are written to the global environment (and will overwrite any existing tables
with the same names). Diplays "Data loaded" if successful.
}
\description{
Loads the HAUL, CATCH, LENGTH, WEIGHT, STATIONS, and SPECIES_CODES tables
from the RREAS survey as currently stored on Dryad. The function also loads a
HAULSTANDARD table with a standardized set of columns including YEAR, MONTH,
JDAY, and lat/lon in decimal degrees. See \code{\link{RREAS_TABLES}} for metadata.
}
\details{
The data tables are stored internally in the package (internet
connection is not required). The maintainers will update the package data
whenever the data on Dryad is updated.

Abundance (including abundance with size limits) and size distributions can
be obtained from this dataset using \code{\link{get_totals}} and \code{\link{get_distributions}}.

The AGE table is not included, so requests for age standardized abundance or
age distributions will be unsuccessful.
}
\examples{
load_trawls()
load_trawls(activestationsonly=FALSE, startyear=1990)

}
\references{
Sakuma, K.M., Field, J.C., Mantua, N.J., Ralston, S., Marinovic,
B.B. and Carrion, C.N. (2016) Anomalous epipelagic micronekton assemblage patterns
in the neritic waters of the California Current in spring 2015 during a
period of extreme ocean conditions. CalCOFI Rep. 57:163-183
}
\seealso{
\code{\link{load_erddap}}, \code{\link{RREAS_TABLES}}
}
\keyword{functions}
