% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sptable_lw}
\alias{sptable_lw}
\title{Species table for RREAS species and species groups (biomass)}
\format{
A data frame with 92 rows and 3 variables:
\describe{
\item{SPECIES}{Species code}
\item{MATURITY}{Maturity code}
\item{NAME}{Name}
}
}
\usage{
sptable_lw
}
\description{
Table of species and species groups for which biomass estimates can be obtained.
}
\keyword{datasets}
