% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sptable_rockfish100}
\alias{sptable_rockfish100}
\title{Species table for rockfishes used in 100 day index}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{SPECIES}{Species code}
\item{MATURITY}{Maturity code}
\item{NAME}{Name (3 letter code)}
\item{MINLEN}{Minimum length (20 mm)}
\item{COMMON}{Common name}
}
}
\usage{
sptable_rockfish100
}
\description{
Species table for rockfishes used in 100 day index
}
\keyword{datasets}
