% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbers.R
\name{get_distributions}
\alias{get_distributions}
\title{Get size, mass, or age distribution data}
\usage{
get_distributions(
  speciestable,
  datasets = "RREAS",
  startyear = 1983,
  what = c("size", "mass", "age"),
  haultable = HAULSTANDARD
)
}
\arguments{
\item{speciestable}{Dataframe containing species information (see Details).}

\item{datasets}{Character vector indicating which dataset(s) to use. Multiple
dataset can be specified. Options are "RREAS", "ADAMS", "PWCC", "NWFSC". If
unspecified, just uses "RREAS".}

\item{startyear}{Start year (default is 1983).}

\item{what}{What type of distribution you want, either "size", "mass", or
"age".}

\item{haultable}{The table of hauls from which to obtain data. Defaults to
HAULSTANDARD. If you do not want to use HAULSTANDARD, you
can specify another table here, but it should have all of the same columns as
HAULSTANDARD.}
}
\value{
A dataframe with haul information, NAME, TOTAL_NO, NMEAS (number
measured), EXP (expansion factor), SP_NO (specimen number), and values for
the requested distribution. If "size" is requested, will include column
STD_LENGTH (mm). If "mass" is requested, will include columns STD_LENGTH and
WEIGHT (g). If "age" is requested, will include columns STD_LENGTH, AGE (days), N100i
(number of 100 day equivalents), and JDAY_DOB (julian date of birth). If size
limits are specified, will include additional columns NMEAS_SIZE (number
measured in the size range), PSIZE (proportion of measured fish in the size
range), and NSIZE (total number in the size range, which is probably what
you want, not TOTAL_NO). If multiple NAME values were present, the data
will be in long format (stacked). If you request data from multiple
datasets, they results will be combined (column SURVEY differentiates
source). The NAME field is converted to a factor so names will plot in the
same order supplied in \code{speciestable}.
}
\description{
Obtains size, mass, or age distribution data for select species for the hauls
in HAULSTANDARD. There will be \emph{multiple} entries per haul. Data from
multiple species can be obtained with a single function call.
}
\details{
Biomass distributions are only available for species with length-weight
regressions, and age distributions are only available for species with
length-age regression. See \code{\link{get_totals}} for more details.

If a haul had no fish, it will appear in the output dataset (with
TOTAL_NO=0). If a haul had fish, but no fish were measured, there will be a
TOTAL_NO>0, NMEAS will be 0, and there will be a single length/mass/age entry
for that haul, which will be the average values used as a substitute. If a
haul has length measurements, there will be multiple entries for that haul,
and haul-level information will be repeated.

The \code{speciestable} should be formatted in the same way as described in
\code{\link{get_totals}}.
}
\examples{
\dontrun{
rockfish100agedist <- get_distributions(sptable_rockfish100, what = "age")
}
}
\seealso{
\code{\link{get_totals}}, \code{\link{get_numbers}}, \code{\link{get_lw_regression}}, \code{\link{get_la_regression}},
\code{\link{age_to_100day}}, \code{\link{sptable}}, \code{\link{sptable_lw}}, \code{\link{sptable_rockfish100}}
}
\keyword{functions}
