% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbers.R
\name{get_distributions}
\alias{get_distributions}
\title{Get size, mass, or age distribution data}
\usage{
get_distributions(
  speciestable,
  datasets = "RREAS",
  startyear = 1983,
  what = c("size", "mass", "age")
)
}
\arguments{
\item{speciestable}{Dataframe containing species information (see Details).}

\item{datasets}{Character vector indicating which dataset(s) to use. Multiple
dataset can be specified. Options are "RREAS","ADAMS","PWCC","NWFSC". If
unspecified, just uses RREAS.}

\item{startyear}{Start year (default is 1983).}

\item{what}{What type of distribution you want, either "size","mass", or
"age".}
}
\value{
A dataframe with haul information, NAME, TOTAL_NO, NMEAS (number
measured), EXP (expansion factor), SP_NO (specimen number) and values for
the requested distribution. If "size" is requested, will include column
STD_LENGTH. If "mass" is requested, will include columns STD_LENGTH and
WEIGHT. If "age" is requested, will include columns STD_LENGTH, AGE, N100i
(number of 100 day equivalents), and JDAY_DOB (date of birth). If size
limits are specified, will include additional columns NMEAS_SIZE (number
measured in the size range), PSIZE (proportion of measured fish in the size
range), and NSIZE (total number in the size range, which is probably what
you want, not TOTAL_NO). If multiple NAME values were present, the data
will be in long format (stacked). If you request data from multiple
datasets, they results will be combined (column SURVEY differentiates
source).
}
\description{
Obtains size, mass, or age distribution data for select species for the hauls
in HAULSTANDARD. There will be \emph{multiple} entries per haul. Data from
multiple species can be obtained with a single function call.
}
\details{
Biomass distributions are only available for species with length-weight
regressions, and age distributions are only available for species with
length-age regression. See \code{\link{get_totals}} for more details.

If a haul had no fish, it will appear in the output dataset (with
TOTAL_NO=0), and the other values will be NA. If a haul had fish, but no fish
were measured, there will be a TOTAL_NO>0, NMEAS will be NA, and there will
be a single length/mass/age entry for that haul, which will be the average
values used as a substitute. If a haul has length measurements, there will be
multiple entries for that haul, and haul-level information will be repeated.

The \code{speciestable} should be formatted in the same way as described in
\code{\link{get_totals}}.
}
\examples{
\dontrun{

}
}
\keyword{functions}
