% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_generation.R
\name{get_logcpueindex}
\alias{get_logcpueindex}
\title{Get CPUE index}
\usage{
get_logcpueindex(df, var = "TOTAL_NO", group = NULL, standardized = TRUE)
}
\arguments{
\item{df}{Data frame output from \code{\link{get_totals}}.}

\item{var}{The variable in \code{df} for which to compute the index, typically "TOTAL_NO"
(default), "BIOMASS", or "N100".}

\item{group}{Optional grouping variable in \code{df} for which to compute separate
indices. Typically "STRATA". Multiple grouping variables are allowed.}

\item{standardized}{Whether or not to compute a scaled index, in addition. Defaults
to TRUE. Scaling is done within groups for each NAME, across years.}
}
\value{
A data frame with columns NAME, YEAR, any grouping variables, and an index.
}
\description{
Given an output table from \code{\link{get_totals}}, calculates the mean log(CPUE+1)
for each YEAR and NAME. Allows an optional grouping variable (e.g. STRATA).
}
\details{
If there are multiple hauls at the same station in a given year, these are
averaged first, and then all stations means in a given year are averaged.

NAs will be filled in for years with no hauls in a given grouping.
}
\examples{
\dontrun{
rockfish100equiv <- get_totals(sptable_rockfish100, what = "100day")
rockfish100index <- get_logcpueindex(rockfish100equiv, var="N100", group="STRATA")
}
}
\keyword{functions}
