% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{get_lw_regression}
\alias{get_lw_regression}
\title{Get length-weight regression}
\usage{
get_lw_regression(species, maturity, plot = F)
}
\arguments{
\item{species}{Species code}

\item{maturity}{Maturity code}

\item{plot}{Logical. Whether to produce a plot (logical). Plots will not be
produced for published regressions, just ones fit using the WEIGHT data.}
}
\value{
A function which takes a vector of lengths (mm) as input, and returns
a vector of masses (g).
}
\description{
Runs a length-weight regression using data in the WEIGHT table and outputs a
function to convert lengths into weights. This function (and the outputed
function) are used internally in \code{\link{get_totals}} and \code{\link{get_distribution}} but
it can be called independently, e.g. if you want to see what the regressions
look like.
}
\details{
This fits a linear regression to log-log data. The outputed conversion
function will do the transformations and backtransformations automatically, so
the inputed lengths should be in the original units (mm).

Length-weight regressions can be done for all species in the WEIGHT table
(run \code{unique(WEIGHT$SPECIES)} for a list) and for all species in
\code{\link{sptable_lw}}.

The rockfishes are divided into 4 groups given in table \code{\link{rflwgroups}}. The
groups are based on general body shape and similarity of regression
intercepts. Within a given group, species with length-weight data are pooled,
and the regression is used for all species in that group, including those
without length-weight data. Thus the species with data act as proxies for
species without data.

For anchovy (209), adults (A) and juveniles (Y) are pooled for the
regression.

For northern lampfish (661) and California lanternfish (669), data from both
species are pooled for the regression. Unknown myctophids (407) will use
this regression.

For octopus (2026), which has weight but no length data, the mean weight is
used.

This function includes and will use published length-weight regressions for
select species not in the WEIGHT table. These species are sardine (562), T.
spinfera (1473, also used by 1791), and E. pacifica (1816, also used by 1472,
1846, 1847, 2829, 2830, 2835, 2849).

If the function is unable to generate a regression for the input species, a
warning will be displayed and the resulting function will return NA for any
length input.
}
\examples{
print(rflwgroups)
\dontrun{
get_lw_regression(species = 209, maturity = "A", plot=T)
get_lw_regression(species = 562, maturity = "A")
}
}
\seealso{
\code{\link{sptable_lw}}, \code{\link{rflwgroups}}
}
\keyword{functions}
