% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numbers.R
\name{get_totals}
\alias{get_totals}
\title{Get total haul-level abundance, biomass, or 100-day standardized abundance}
\usage{
get_totals(
  speciestable,
  datasets = "RREAS",
  startyear = 1983,
  what = c("abundance", "biomass", "100day")
)
}
\arguments{
\item{speciestable}{Dataframe containing species information (see Details).}

\item{datasets}{Character vector indicating which dataset(s) to use. Multiple dataset can
be specified. Options are "RREAS","ADAMS","PWCC","NWFSC". If unspecified, just uses RREAS.}

\item{startyear}{Start year (default is 1983).}

\item{what}{What totals you want, either "abundance","biomass", or "100day".}
}
\value{
A dataframe with haul information, NAME, and totals. If "abundance"
is requested, will include column TOTAL_NO. If "biomass" is requested, will
include columns TOTAL_NO, NMEAS (number measured), and BIOMASS. If "100day"
is requested, will include columns TOTAL_NO, NMEAS (number measured), and
N100. If size limits are specified, will include additional columns
NMEAS_SIZE (number measured in the size range), and NSIZE (total number in
the size range, which is probably what you want, not TOTAL_NO). If multiple
NAME values were present, the data will be in long format (stacked). If you
request data from multiple datasets, they results will be combined (column
SURVEY differentiates source).
}
\description{
Obtains total haul-level abundance, biomass, or 100-day standardized
abundance for select species the hauls in HAULSTANDARD. There will be a
single entry per haul. Data from multiple species can be obtained with a
single function call (see Details).
}
\details{
Note that if a station was sampled but the species requested does not appear
in the CATCH table for that haul, it will appear as a 0. It your
responsibility to know when species were or were not being recorded, and thus
whether the 0 represents 'absent' or 'not counted'. Some info can be found in the
SPECIES_CODES table.

Length data are required for "biomass", "100day", and length-specific
abundances. If length data are available for a species, but not available for
a particular haul where the species was present, mean length values will be
used. If available, the mean for the same CRUISE and REGION will be used,
followed by CRUISE and STRATA, then CRUISE, then the global mean. In the
output table, if TOTAL_NO>0 but NMEAS is 0, this indicates that length data
were unavailable and means were used. If size limits are specified, the mean
proportion of fish in the length range will also be used for hauls missing
length data, with the same rank ordering of available mean values.

Biomass is only available for species with length-weight regressions. Species
available for biomass estimates are listed in \code{\link{sptable_lw}}. See
\code{\link{get_lw_regression}} (used internally) for more details.

100 day standardized abundance is only available for species with length-age
regression. This includes the rockfish species listed in
\code{\link{sptable_rockfish100}}, hake (382) and lingcod (448). See
\code{\link{get_la_regression}} (used internally) for more details. Note that all
species will use the same age-abundance conversion.  See \code{\link{age_to_100day}}
(used internally) for more details.

\strong{Formatting the species table}

The structure of the \code{speciestable} determines which species will be
pulled whether multiple species codes should be added together under one
category (e.g. total rockfish). It needs to be in a particular format.

The package contains several tables already in the correct format, which can
be passed directly to the \code{speciestable} argument, and can be subsetted.
\code{\link{sptable}} contains a suite of common species for which abundance indices
are typically calculated. \code{\link{sptable_lw}} contains species for which biomass
can be obtained. \code{\link{sptable_rockfish100}} contains the rockfish species used
to generate the 100 day index, and specifies a min length of 20.

The \code{speciestable} should be a data frame with the following columns:
\describe{
\item{SPECIES}{Species code, as in table SPECIES_CODES}
\item{MATURITY}{Maturity code, either "Y", "A", or "U"}
\item{NAME}{Custom name that will be assigned in the output table.
Rows with the same NAME will be combined in the output}
\item{MINLEN}{(Optional) Minimum length in mm (>=).}
\item{MAXLEN}{(Optional) Maximum length in mm (<).}
}

Specifying min and/or max lengths is optional. If columns are omitted or values are NA, min
length defaults to 0 and max length to Inf (in other words, no constraints).

If multiple datasets are specified, these will be combined in the output table (column SURVEY
will indicate origin).
}
\examples{
\dontrun{
krillsp=data.frame(SPECIES=c(1473,1816),
  MATURITY=c("U","U"),
  NAME=c("T. spinifera","E. pacifica"))
krillabundance=get_totals(speciestable=krillsp, what="abundance")

somebiomasses=get_totals(speciestable=sptable, what="biomass")

rockfish100day=get_totals(speciestable=sptable_rockfish100, what="100day",
  datasets=c("RREAS","ADAMS","PWCC","NWFSC"))
}
}
\keyword{functions}
