% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_mdb}
\alias{load_mdb}
\title{Load RREAS Trawl Data from MS Access Database}
\usage{
load_mdb(
  mdb_path,
  atsea_path = NULL,
  datasets = "RREAS",
  krill_len_path = NULL,
  activestationsonly = TRUE,
  stdtimeperiodonly = TRUE,
  startyear = 1983,
  devswitch = FALSE
)
}
\arguments{
\item{mdb_path}{File path to the juv_cruise mdb file (required).}

\item{atsea_path}{File path to "atsea.mdb" containing the current year's data (optional).
Data will be appended to the RREAS tables.}

\item{datasets}{Character vector indicating which dataset(s) to load. Multiple dataset can
be specified. Options are "RREAS", "ADAMS", "PWCC", "NWFSC". If unspecified, loads just "RREAS".}

\item{krill_len_path}{File path to "krill_lengths.csv" (optional, unless you later want
to get krill biomass).
\emph{This argument will be removed once the krill lengths are added to the database.}}

\item{activestationsonly}{Logical. Include only active stations in HAULSTANDARD. Defaults to TRUE.
If FALSE, the ACTIVE column can always be used to subset later.}

\item{stdtimeperiodonly}{Logical. Include only surveys performed during the standard sampling
time period in HAULSTANDARD (exclude early surveys). Defaults to TRUE.}

\item{startyear}{Start year (default is 1983).}

\item{devswitch}{Patch for developer use only.}
}
\value{
Tables are written to the global environment (and will overwrite any existing tables
with the same names). Diplays "Data loaded" if successful.
}
\description{
Loads the HAUL, CATCH, and LENGTH tables from one or more surveys in the RREAS database,
along with the AGE, WEIGHT, STATIONS, and SPECIES_CODES tables from RREAS. Also creates and loads a
HAULSTANDARD table for each survey containing only standard stations and with a standardized set of
columns including YEAR, MONTH, JDAY, and lat/lon in decimal degrees. HAULSTANDARD tables all have
the same format and can be stacked with \code{rbind}.
\strong{This function requires a local copy of the RREAS MS Access Database.}
}
\details{
If the "NWFSC" dataset is requested, then the HAULSTANDARD_NWFSC table will
contain the NWFSC stations sampled in the NWFSC survey, the RREAS stations
sampled in the NWFSC survey, and the NWFSC stations sampled in the RREAS
survey.

If a path to "atsea.mdb" is given, containing the current year's data, it
will be appended to the RREAS tables.
}
\examples{
\dontrun{
load_mdb(mdb_path = "insert_filepath/juv_cruise_backup27JAN21.mdb",
  krill_len_path = "insert_filepath/krill_lengths.csv",
  datasets = c("RREAS","ADAMS","PWCC","NWFSC"),
  activestationsonly = F)
}
}
\references{
Sakuma, K.M., Field, J.C., Mantua, N.J., Ralston, S., Marinovic,
B.B. and Carrion, C.N. (2016) Anomalous epipelagic micronekton assemblage patterns
in the neritic waters of the California Current in spring 2015 during a
period of extreme ocean conditions. CalCOFI Rep. 57:163-183
}
\keyword{functions}
