% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rflwgroups}
\alias{rflwgroups}
\title{Rockfish length-weight groupings}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{SPECIES}{Species code}
\item{COMMON_NAME}{Common name}
\item{SCI_NAME}{Scientific name}
\item{RFGROUP}{Group number}
\item{SPECIES_GROUP}{Rockfish}
\item{NMEAS}{Number measured}
}
}
\usage{
rflwgroups
}
\description{
Length-weight groupings for all rockfish species, specifying which species
should be pooled for regressions and used as proxies for other species.
There are currently 4 groupings based on general body shape.
}
\keyword{datasets}
