% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RREAS_TABLES}
\alias{RREAS_TABLES}
\alias{STATIONS_RREAS}
\alias{SPECIES_CODES_RREAS}
\alias{CATCH_ERDDAP}
\alias{CATCH_RREAS}
\alias{HAUL_ERDDAP}
\alias{HAUL_RREAS}
\alias{HAULSTANDARD_ERDDAP}
\alias{HAULSTANDARD_RREAS}
\alias{WEIGHT_RREAS}
\alias{CTD_HEADER}
\alias{CTD_CAST}
\alias{LENGTH_ERDDAP}
\alias{LENGTH_RREAS}
\title{RREAS data tables}
\format{
Data frames with some combination of the following fields
\describe{
\item{CRUISE}{Cruise ID}
\item{HAUL_NO}{Haul number}
\item{VESSEL}{Survey vessel code}
\item{STATION}{Station code}
\item{HAUL_DATE}{Haul date and time}
\item{YEAR}{Survey year}
\item{MONTH}{Haul month}
\item{JDAY}{Haul julian day}
\item{NET_IN_LATDD}{Latitude at which net went in water (decimal degrees)}
\item{NET_IN_LONDD}{Longitude at which net went in water (decimal degrees)}
\item{LATDD}{Latitude of station (decimal degrees)}
\item{LONDD}{Longitude of station (decimal degrees)}
\item{BOTTOM_DEPTH}{Bottom depth of haul (m)}
\item{STATION_BOTTOM_DEPTH}{Bottom depth of station (m)}
\item{STRATA}{Survey region (C:Core, NC:North core, SC: South core, N: North, S: South)}
\item{AREA}{Survey area (transect or subregion)}
\item{SPECIES}{Species code}
\item{MATURITY}{Maturity code}
\item{COMMON_NAME}{Common name}
\item{SCI_NAME}{Scientific name}
\item{TOTAL_NO}{Number of individuals caught}
\item{STD_LENGTH}{Standard length (mm)}
}

An object of class \code{data.frame} with 291 rows and 8 columns.

An object of class \code{data.frame} with 213 rows and 6 columns.

An object of class \code{data.frame} with 3274 rows and 14 columns.

An object of class \code{data.frame} with 5129 rows and 35 columns.

An object of class \code{data.frame} with 3274 rows and 17 columns.

An object of class \code{data.frame} with 4560 rows and 17 columns.

An object of class \code{data.frame} with 52604 rows and 5 columns.

An object of class \code{data.frame} with 75959 rows and 6 columns.

An object of class \code{data.frame} with 231494 rows and 5 columns.

An object of class \code{data.frame} with 335125 rows and 6 columns.

An object of class \code{data.frame} with 1015 rows and 6 columns.
}
\source{
\url{https://oceanview.pfeg.noaa.gov/erddap/tabledap/FED_Rockfish_Catch.html}
\url{https://oceanview.pfeg.noaa.gov/erddap/tabledap/FED_Rockfish_Length.html}
}
\usage{
STATIONS_RREAS

SPECIES_CODES_RREAS

HAUL_ERDDAP

HAUL_RREAS

HAULSTANDARD_ERDDAP

HAULSTANDARD_RREAS

CATCH_ERDDAP

CATCH_RREAS

LENGTH_ERDDAP

LENGTH_RREAS

WEIGHT_RREAS
}
\description{
Meta-data for RREAS data tables. Additional details can be found in the data publication.
}
\details{
Data from ERDDAP are reformatted as relational tables matching the
format of the tables in the database. Data on ERDDAP begins in 1990 and
ends about 2 years from the present. The ERDDAP version contains additional
SPECIES/MATURITY categories 1472/T (total krill) and 1940/T (total
rockfish).
}
\references{
Sakuma, K.M., Field, J.C., Mantua, N.J., Ralston, S., Marinovic,
B.B. and Carrion, C.N. (2016) Anomalous epipelagic micronekton assemblage patterns
in the neritic waters of the California Current in spring 2015 during a
period of extreme ocean conditions. CalCOFI Rep. 57:163-183
}
\keyword{datasets}
