% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sptable}
\alias{sptable}
\title{Species table for common RREAS species and species groups}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{SPECIES}{Species code}
\item{MATURITY}{Maturity code}
\item{NAME}{Name}
}
}
\usage{
sptable
}
\description{
Table of species and species groups for which indices are commonly
reported in ecosystem assessments.
}
\keyword{datasets}
